ai-services/requirements.txt
numpy==2.2.2
pillow==11.1.0
websockets==14.2
pytest==8.3.4
